Move corrupt member test to shared decoder suite
